refactor(sources): modularize scraping handlers
